+++ CHANGES.md
@@ -6,6 +6,7 @@
 - **ENH: Use `pyproject.toml` instead of `setup.py`**
 - FIX: Fix too broad exception in case of `geopandas >= 1.0`
 - FIX: Fix deprecation warning for `get_nodata_value_from_dtype` in `rasters_rio`
+- OPTIM: Compute the spatial index by default in `vectors.read` (set `vectors.read(..., compute_sindex=False)` if you don't want to compute them)
 
 ## 1.44.1 (2024-12-12)